requirements.txt
fastapi==0.68.2
pydantic==1.10.12
uvicorn==0.15.0
Jinja2==3.0.1
SQLAlchemy==1.4.22
alembic==1.6.5
psycopg2==2.9.7
pydantic[email]
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.24.1